requirements.txt
streamlit
ollama
SpeechRecognition
gTTS
pygame
pydub
pyaudio
psycopg2-binary
python-dotenv
redis
duckduckgo-search
crawl4ai
playwright
